Fix for BlocksStructure terminating early
Changes to BlocksStructure had caused the excited states generator to terminate early
Also removed use of int64_t on mac in arpack
And removed use of prediction vecotr for excited states, as this currently does not
generate the correct prediction vector

+++ arpack_interface.hpp
@@ -20,7 +20,7 @@ namespace arpack {
   typedef long long arpack_int; 
 #else
   /** Else use normal int type */
-  typedef int64_t arpack_int;
+  typedef int arpack_int;
 #endif
 
   /** external arpack function */
